Learn Linux In 5 Days
In this eBook, you will learn the most important concepts and commands, and be guided step-by-step through several practical
and real-world examples.

The most important material is condensed into five sections, each designed to be consumed in a day.
Here is what you will learn: How to gain access to a Linux server and what Linux distribution is,
What software is needed and what SSH is, The file system layout of Linux systems and where to find
programs and documentation, Basic Linux commands you'll use the most often, and much, much more!