Postgres
Although the number of database systems on the market is constantly climbing,
the Postgres database continues to be an under-championed, reliable option for many scenarios.
Its open-source nature, cross-platform flexibility, and wide selection of additional tools are
just some of the strengths that make this system a go-to for any data management need, enterprise or otherwise.
In this book the author guides you from installing Postgres, to creating your first database,
to using Postgres in a .NET application with ADO .NET and Entity Framework.